Move settings store logic to settings directory
