+++ src/app/effects/index.ts
@@ -4,4 +4,4 @@ export * from './getting-started.effects';
 export * from './login.effects';
 export * from './model.effects';
 export * from './router.effects';
-export * from './settings.effects';
+export * from '../settings/store/settings.effects';
