+++ src/app/reducers/index.ts
@@ -6,7 +6,7 @@ import * as fromAuth from './auth.reducer';
 import * as fromCheckIn from './check-in.reducer';
 import * as fromGettingStarted from './getting-started.reducer';
 import * as fromModel from './model.reducer';
-import * as fromSettings from './settings.reducer';
+import * as fromSettings from '../settings/store/settings.reducer';
 
 export interface State {
   app: fromApp.State;
