+++ src/app/settings/containers/roles/roles.component.ts
@@ -3,7 +3,7 @@ import { Store } from '@ngrx/store';
 import { Observable, Subscription } from 'rxjs';
 import { map } from 'rxjs/internal/operators';
 
-import * as SettingsActions from '../../../actions/settings.actions';
+import * as SettingsActions from '../../store/settings.actions';
 import { canSelectRole, getRoleOptions, isLastOwner } from '../../../functions';
 import { ColumnOptions, User } from '../../../models';
 import { State } from '../../../reducers';
